requirements.txt
requests
pyyaml
mutagen==1.45.1
